FEAT : Game Balancing 1

+++ Assets/02.Script/Character/EnemyBase.cs
@@ -123,7 +123,7 @@ public class EnemyBase : MonoBehaviour
         // 1000미만 => 그냥
         // 1000000미만 => 1000으로 나누고 K
         // 그 외 => 1000000으로 나누고 M
-        if(dmg < 1000) instantFloatingText.text.text = dmg.ToString();
+        if(dmg < 1000) instantFloatingText.text.text = ((int)dmg).ToString();
         else if(dmg < 1000000) instantFloatingText.text.text = $"{dmg / 1000}K";
         else instantFloatingText.text.text = $"{dmg / 1000000}M";
